Move debugger files from platform folder
