+++ package.nls.json
@@ -408,7 +408,6 @@
     "DataScience.cellStopOnErrorMessage": "Cell was canceled due to an error in a previous cell.",
     "DataScience.instructionComments": "# To add a new cell, type '{0}'\n# To add a new markdown cell, type '{0} [markdown]'\n",
     "DataScience.scrollToCellTitleFormatMessage": "Go to [{0}]",
-    "DataScience.remoteDebuggerNotSupported": "Debugging while attached to a remote server is not currently supported.",
     "DataScience.save": "Save notebook",
     "DataScience.nativeEditorTitle": "Notebook Editor",
     "DataScience.untitledNotebookFileName": "Untitled",
