+++ src/kernels/debugger/types.ts
@@ -1,6 +1,9 @@
 // Copyright (c) Microsoft Corporation. All rights reserved.
 // Licensed under the MIT License.
+'use strict';
 
+import { DebugProtocol } from 'vscode-debugprotocol';
+import { IDebugService } from '../../platform/common/application/types';
 import {
     DebugAdapter,
     DebugConfiguration,
@@ -10,7 +13,52 @@ import {
     NotebookCell,
     NotebookDocument
 } from 'vscode';
-import { DebugProtocol } from 'vscode-debugprotocol';
+
+export interface ISourceMapMapping {
+    line: number;
+    endLine: number;
+    runtimeSource: { path: string };
+    runtimeLine: number;
+}
+
+export interface ISourceMapRequest {
+    source: { path: string };
+    pydevdSourceMaps: ISourceMapMapping[];
+}
+
+export const IJupyterDebugService = Symbol('IJupyterDebugService');
+export interface IJupyterDebugService extends IDebugService {
+    /**
+     * Event fired when a breakpoint is hit (debugger has stopped)
+     */
+    readonly onBreakpointHit: Event<void>;
+    /**
+     * Start debugging a notebook cell.
+     * @param nameOrConfiguration Either the name of a debug or compound configuration or a [DebugConfiguration](#DebugConfiguration) object.
+     * @return A thenable that resolves when debugging could be successfully started.
+     */
+    startRunByLine(config: DebugConfiguration): Thenable<boolean>;
+    /**
+     * Gets the current stack frame for the current thread
+     */
+    getStack(): Promise<DebugProtocol.StackFrame[]>;
+    /**
+     * Steps the current thread. Returns after the request is sent. Wait for onBreakpointHit or onDidTerminateDebugSession to determine when done.
+     */
+    step(): Promise<void>;
+    /**
+     * Runs the current thread. Will keep running until a breakpoint or end of session.
+     */
+    continue(): Promise<void>;
+    /**
+     * Force a request for variables. DebugAdapterTrackers can listen for the results.
+     */
+    requestVariables(): Promise<void>;
+    /**
+     * Stop debugging
+     */
+    stop(): void;
+}
 
 export interface IKernelDebugAdapter extends DebugAdapter {
     stepIn(threadId: number): Thenable<DebugProtocol.StepInResponse['body']>;
