+++ src/notebooks/debugger/debuggingManager.ts
@@ -17,25 +17,25 @@ import {
     EventEmitter,
     NotebookEditor
 } from 'vscode';
-import * as path from '../../../vscode-path/path';
-import { IKernel, IKernelProvider } from '../../../../kernels/types';
-import { IConfigurationService, IDisposable } from '../../../common/types';
-import { KernelDebugAdapter } from '../kernelDebugAdapter';
-import { IExtensionSingleActivationService } from '../../../activation/types';
-import { ContextKey } from '../../../common/contextKey';
-import { IApplicationShell, ICommandManager, IVSCodeNotebook } from '../../../common/application/types';
-import { traceError, traceInfo, traceInfoIfCI } from '../../../logging';
-import { DataScience } from '../../../common/utils/localize';
-import { Commands as DSCommands, EditorContexts } from '../../../../webviews/webview-side/common/constants';
-import { IPlatformService } from '../../../common/platform/types';
-import { IDebuggingManager, IKernelDebugAdapterConfig, KernelDebugMode } from '../../types';
-import { DebuggingTelemetry, pythonKernelDebugAdapter } from '../../constants';
-import { sendTelemetryEvent } from '../../../../telemetry';
-import { DebugCellController } from './debugCellControllers';
-import { assertIsDebugConfig, IpykernelCheckResult, isUsingIpykernel6OrLater } from '../helper';
-import { Debugger } from '../debugger';
-import { INotebookControllerManager } from '../../../../notebooks/types';
+import * as path from '../../platform/vscode-path/path';
+import { IKernel, IKernelProvider } from '../../kernels/types';
+import { IConfigurationService, IDisposable } from '../../platform/common/types';
+import { Commands as DSCommands, EditorContexts } from '../../webviews/webview-side/common/constants';
+import { IExtensionSingleActivationService } from '../../platform/activation/types';
+import { ContextKey } from '../../platform/common/contextKey';
+import { Debugger } from './debugger';
 import { RunByLineController } from './runByLineController';
+import { INotebookControllerManager } from '../types';
+import { IApplicationShell, ICommandManager, IVSCodeNotebook } from '../../platform/common/application/types';
+import { IPlatformService } from '../../platform/common/platform/types';
+import { DebuggingTelemetry, pythonKernelDebugAdapter } from '../../kernels/debugger/constants';
+import { sendTelemetryEvent } from '../../telemetry';
+import { traceError, traceInfo, traceInfoIfCI } from '../../platform/logging';
+import { DataScience } from '../../platform/common/utils/localize';
+import { DebugCellController } from './debugCellControllers';
+import { KernelDebugAdapter } from './kernelDebugAdapter';
+import { assertIsDebugConfig, IpykernelCheckResult, isUsingIpykernel6OrLater } from './helper';
+import { IDebuggingManager, IKernelDebugAdapterConfig, KernelDebugMode } from '../../kernels/debugger/types';
 
 /**
  * The DebuggingManager maintains the mapping between notebook documents and debug sessions.
