+++ src/platform/serviceRegistry.node.ts
@@ -32,9 +32,7 @@ import { ExtensionRecommendationService } from './common/extensionRecommendation
 import { GlobalActivation } from './common/globalActivation';
 import { PreReleaseChecker } from './common/prereleaseChecker.node';
 import { IConfigurationService, IDataScienceCommandListener, IExtensionContext } from './common/types';
-import { DebugLocationTrackerFactory } from './debugger/debugLocationTrackerFactory.node';
-import { DebuggingManager } from './debugger/jupyter/notebook/debuggingManager';
-import { IDebugLocationTracker, IDebuggingManager } from './debugger/types';
+import { DebugLocationTrackerFactory } from '../kernels/debugger/debugLocationTrackerFactory.node';
 import { DataScienceErrorHandler } from './errors/errorHandler';
 import { IDataScienceErrorHandler } from './errors/types';
 import { ExportBase } from './export/exportBase.node';
@@ -64,6 +62,8 @@ import { WorkspaceService } from './common/application/workspace.node';
 import { ExtensionSideRenderer, IExtensionSideRenderer } from '../webviews/extension-side/renderer';
 import { OutputCommandListener } from './logging/outputCommandListener';
 import { ExportUtilBase } from './export/exportUtil';
+import { DebuggingManager } from '../notebooks/debugger/debuggingManager';
+import { IDebuggingManager, IDebugLocationTracker } from '../kernels/debugger/types';
 
 export function registerTypes(context: IExtensionContext, serviceManager: IServiceManager, isDevMode: boolean) {
     serviceManager.addSingleton<FileSystem>(FileSystem, FileSystem);
