+++ src/platform/serviceRegistry.web.ts
@@ -28,8 +28,6 @@ import { GlobalActivation } from './common/globalActivation';
 import { IExtensionSingleActivationService } from './activation/types';
 import { ExtensionSideRenderer, IExtensionSideRenderer } from '../webviews/extension-side/renderer';
 import { OutputCommandListener } from './logging/outputCommandListener';
-import { IDebuggingManager } from './debugger/types';
-import { DebuggingManager } from './debugger/jupyter/notebook/debuggingManager';
 import { ExportDialog } from './export/exportDialog';
 import { ExportFormat, IExport, IExportBase, IExportDialog, IFileConverter, INbConvertExport } from './export/types';
 import { FileConverter } from './export/fileConverter';
@@ -46,6 +44,8 @@ import { NotebookWatcher } from '../webviews/extension-side/variablesView/notebo
 import { DataViewerFactory } from '../webviews/extension-side/dataviewer/dataViewerFactory';
 import { IDataViewerFactory } from '../webviews/extension-side/dataviewer/types';
 import { INotebookWatcher } from '../webviews/extension-side/variablesView/types';
+import { DebuggingManager } from '../notebooks/debugger/debuggingManager';
+import { IDebuggingManager } from '../kernels/debugger/types';
 
 export function registerTypes(context: IExtensionContext, serviceManager: IServiceManager, isDevMode: boolean) {
     serviceManager.addSingleton<IFileSystem>(IFileSystem, FileSystem);
