+++ src/telemetry/index.ts
@@ -23,7 +23,7 @@ import { ResourceSpecificTelemetryProperties } from './types';
 import { CheckboxState, EventName, PlatformErrors, SliceOperationSource } from './constants';
 import { noop } from '../platform/common/utils/misc';
 import { isPromise } from 'rxjs/internal-compatibility';
-import { DebuggingTelemetry } from '../platform/debugger/constants';
+import { DebuggingTelemetry } from '../kernels/debugger/constants';
 import { EnvironmentType } from '../platform/pythonEnvironments/info';
 import { TelemetryErrorProperties, ErrorCategory } from '../platform/errors/types';
 import { ExportFormat } from '../platform/export/types';
