+++ src/test/datascience/debugger.vscode.test.ts
@@ -25,7 +25,6 @@ import {
 } from './notebook/helper.node';
 import { ITestVariableViewProvider } from './variableView/variableViewTestInterfaces';
 import { traceInfo } from '../../platform/logging';
-import { IDebuggingManager } from '../../platform/debugger/types';
 import { assert } from 'chai';
 import { debug } from 'vscode';
 import { ITestWebviewHost } from './testInterfaces';
@@ -33,6 +32,7 @@ import { DebugProtocol } from 'vscode-debugprotocol';
 import { waitForVariablesToMatch } from './variableView/variableViewHelpers';
 import { Commands } from '../../platform/common/constants';
 import { IVariableViewProvider } from '../../webviews/extension-side/variablesView/types';
+import { IDebuggingManager } from '../../kernels/debugger/types';
 
 suite('VSCode Notebook - Run By Line', function () {
     let api: IExtensionTestApi;
