+++ src/test/datascience/editor-integration/codelensprovider.unit.test.ts
@@ -17,7 +17,7 @@ import { IConfigurationService, IWatchableJupyterSettings } from '../../../platf
 import { DataScienceCodeLensProvider } from '../../../interactive-window/editor-integration/codelensprovider';
 import { IServiceContainer } from '../../../platform/ioc/types';
 import { ICodeWatcher, IDataScienceCodeLensProvider } from '../../../interactive-window/editor-integration/types';
-import { IDebugLocationTracker } from '../../../platform/debugger/types';
+import { IDebugLocationTracker } from '../../../kernels/debugger/types';
 
 // eslint-disable-next-line
 suite('DataScienceCodeLensProvider Unit Tests', () => {
