+++ src/test/datascience/editor-integration/codewatcher.unit.test.ts
@@ -42,9 +42,9 @@ import { InteractiveCellResultError } from '../../../platform/errors/interactive
 import { ICodeWatcher, IGeneratedCodeStorageFactory } from '../../../interactive-window/editor-integration/types';
 import { IInteractiveWindowProvider, IInteractiveWindow } from '../../../interactive-window/types';
 import { Commands, EditorContexts } from '../../../platform/common/constants';
-import { IDebugLocationTracker } from '../../../platform/debugger/types';
 import { IDataScienceErrorHandler } from '../../../platform/errors/types';
 import { SystemVariables } from '../../../platform/common/variables/systemVariables.node';
+import { IDebugLocationTracker } from '../../../kernels/debugger/types';
 
 /* eslint-disable @typescript-eslint/no-explicit-any */
